Process header images and avatars with Sidekiq
Should help with occasional admin timeouts

Extract an ApplicationUploader class with common configs
