+++ Gemfile
@@ -24,6 +24,7 @@ gem "responders"
 gem "textacular"
 
 gem "carrierwave"
+gem "carrierwave_backgrounder", github: "hyp3rventures/carrierwave_backgrounder" # Forked version for mime-types dependency upgrade
 gem "commonmarker"
 gem "devise"
 gem "fog-aws"
