+++ app/uploaders/application_uploader.rb
@@ -1,0 +1,20 @@
+class ApplicationUploader < CarrierWave::Uploader::Base
+  # Include RMagick or MiniMagick support:
+  # include CarrierWave::RMagick
+  include CarrierWave::MiniMagick
+
+  # Override the directory where uploaded files will be stored.
+  # This is a sensible default for uploaders that are meant to be mounted:
+  def store_dir
+    "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
+  end
+
+  # Add a white list of extensions which are allowed to be uploaded.
+  # For images you might use something like this:
+  def extension_whitelist
+    %w[jpg jpeg gif png]
+  end
+
+  # Set up CarrierWave Backgrounder
+  include ::CarrierWave::Backgrounder::Delay
+end
