+++ app/uploaders/avatar_uploader.rb
@@ -1,16 +1,5 @@
-class AvatarUploader < CarrierWave::Uploader::Base
-
+class AvatarUploader < ApplicationUploader
   include Gravatarify::Helper
-
-  # Include RMagick or MiniMagick support:
-  # include CarrierWave::RMagick
-  include CarrierWave::MiniMagick
-
-  # Override the directory where uploaded files will be stored.
-  # This is a sensible default for uploaders that are meant to be mounted:
-  def store_dir
-    "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
-  end
 
   # Provide a default URL as a default if there hasn't been a file uploaded:
   def default_url(*_args)
@@ -24,28 +13,8 @@ class AvatarUploader < CarrierWave::Uploader::Base
     end
   end
 
-  # Process files as they are uploaded:
-  # process scale: [200, 300]
-  #
-  # def scale(width, height)
-  #   # do something
-  # end
-
   # Create different versions of your uploaded files:
   version :thumb do
     process resize_to_fill: [100, 100]
   end
-
-  # Add a white list of extensions which are allowed to be uploaded.
-  # For images you might use something like this:
-  def extension_whitelist
-    %w(jpg jpeg gif png)
-  end
-
-  # Override the filename of the uploaded files:
-  # Avoid using model.id or version_name here, see uploader/store.rb for details.
-  # def filename
-  #   "something.jpg" if original_filename
-  # end
-
 end
