+++ app/uploaders/header_image_uploader.rb
@@ -1,21 +1,4 @@
-class HeaderImageUploader < CarrierWave::Uploader::Base
-  # Include RMagick or MiniMagick support:
-  # include CarrierWave::RMagick
-  include CarrierWave::MiniMagick
-
-  # Override the directory where uploaded files will be stored.
-  # This is a sensible default for uploaders that are meant to be mounted:
-  def store_dir
-    "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
-  end
-
-  # Process files as they are uploaded:
-  # process scale: [200, 300]
-  #
-  # def scale(width, height)
-  #   # do something
-  # end
-
+class HeaderImageUploader < ApplicationUploader
   # Create different versions of your uploaded files:
   version :thumb do
     process resize_to_limit: [100, 100]
@@ -32,16 +15,4 @@ class HeaderImageUploader < CarrierWave::Uploader::Base
   version :content_card do
     process resize_to_limit: [659 * 2, nil]
   end
-
-  # Add a white list of extensions which are allowed to be uploaded.
-  # For images you might use something like this:
-  def extension_whitelist
-    %w[jpg jpeg gif png]
-  end
-
-  # Override the filename of the uploaded files:
-  # Avoid using model.id or version_name here, see uploader/store.rb for details.
-  # def filename
-  #   "something.jpg" if original_filename
-  # end
 end
