+++ config/initializers/carrierwave_backgrounder.rb
@@ -1,0 +1,3 @@
+CarrierWave::Backgrounder.configure do |c|
+  c.backend :sidekiq, queue: :default
+end
